feat: remove assistant v1 (#2324)
