+++ src/navigation/TabNavigator.tsx
@@ -7,8 +7,7 @@ import {
 import ThemeText from '@atb/components/text';
 import {LocationWithMetadata} from '@atb/favorites/types';
 import {usePreferenceItems} from '@atb/preferences';
-import Assistant_v2 from '@atb/screens/Assistant_v2/';
-import Assistant from '@atb/screens/Assistant/';
+import Assistant from '@atb/screens/Assistant';
 import NearbyScreen, {NearbyStackParams} from '@atb/screens/Nearby';
 import ProfileScreen, {ProfileStackParams} from '@atb/screens/Profile';
 import TicketingScreen from '@atb/screens/Ticketing';
@@ -44,8 +43,7 @@ const Tab = createBottomTabNavigator<TabNavigatorParams>();
 const NavigationRoot = () => {
   const {theme} = useTheme();
   const {t} = useTranslation();
-  const {startScreen, useExperimentalTripSearch, newDepartures} =
-    usePreferenceItems();
+  const {startScreen, newDepartures} = usePreferenceItems();
   const lineHeight = theme.typography.body__secondary.fontSize.valueOf();
 
   return (
@@ -63,7 +61,7 @@ const NavigationRoot = () => {
     >
       <Tab.Screen
         name="Assistant"
-        component={useExperimentalTripSearch ? Assistant_v2 : Assistant}
+        component={Assistant}
         options={tabSettings(
           t(dictionary.navigation.assistant),
           t(dictionary.navigation.assistant),
