+++ src/screens/Assistant/types.tsx
@@ -1,5 +1,9 @@
+import {TripPattern} from '@atb/api/types/trips';
+
 export type SearchStateType =
   | 'idle'
   | 'searching'
   | 'search-success'
   | 'search-empty-result';
+
+export type TripPatternWithKey = TripPattern & {key: string};
