--- src/screens/Assistant_v2/types.tsx
@@ -1,9 +1,0 @@
-import {TripPattern} from '@atb/api/types/trips';
-
-export type SearchStateType =
-  | 'idle'
-  | 'searching'
-  | 'search-success'
-  | 'search-empty-result';
-
-export type TripPatternWithKey = TripPattern & {key: string};
